Move file-tree-utils.ts to file-system/controllers
