+++ src/file-system/controllers/file-tree-utils.ts
@@ -1,4 +1,4 @@
-import type { FileEntry } from "../file-system/models/app";
+import type { FileEntry } from "../models/app";
 
 export function sortFileEntries(entries: FileEntry[]): FileEntry[] {
   return entries.sort((a, b) => {
