+++ src/file-system/controllers/store.ts
@@ -12,14 +12,6 @@ import { useBufferStore } from "../../stores/buffer-store";
 import { useGitStore } from "../../stores/git-store";
 import { useProjectStore } from "../../stores/project-store";
 import { useRecentFoldersStore } from "../../stores/recent-folders-store";
-import {
-  addFileToTree,
-  collapseAllFolders,
-  findFileInTree,
-  removeFileFromTree,
-  sortFileEntries,
-  updateFileInTree,
-} from "../../utils/file-tree-utils";
 import { getGitStatus } from "../../utils/git";
 import { isDiffFile, parseRawDiffContent } from "../../utils/git-diff-parser";
 import type { FileEntry } from "../models/app";
@@ -31,6 +23,14 @@ import {
   readDirectoryContents,
   readFileContent,
 } from "./file-operations";
+import {
+  addFileToTree,
+  collapseAllFolders,
+  findFileInTree,
+  removeFileFromTree,
+  sortFileEntries,
+  updateFileInTree,
+} from "./file-tree-utils";
 import { getFilenameFromPath, getRootPath, isImageFile, isSQLiteFile } from "./file-utils";
 import { useFileWatcherStore } from "./file-watcher-store";
 import { openFolder, readDirectory } from "./platform";
